requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.20.0
loguru==0.6.0
Django==5.0.4
djangorestframework==3.15.2
django-filter==24.2